closes #62
previously, custom.py was being compiled during setup tools into a .pyc (which shouldn't be a part of the experiment setup because it might not be binary compatible across machines).  now that custom.py is stored as custom.py.txt this is no longer an issue and we shouldn't delete the .pyc file.