Remove using_sandbox from config
Instead, whether sandbox mode is in use gets passed as an argument to
the server subprocess and then down to the flask app. Seems to be
working but there are a bunch of files I changed here so some more eyes
would be good.

Right now shell just starts in sandbox automatically. We could change
that and turn the"always_launch_in_sandbox" config option into a
"launch_in_sandbox_or_live" type option, or just remove the option.