create newline on keyboard interrupt.
`cmd2`'s `_cmdloop` only exits when `stop` internally evaluates to True. unless there has been a keyboardinterrupt. So if
stop is False yet we got here then a keyboard interrupt happened. unfortunately we can't simply listen for
a KeyboardInterrupt here because cmd2's `_cmdloop` eats exceptions with a `return` in a `finally` block. so we have to fall back on checking this stop value. we also have to use `_cmdloop()` instead of `cmdloop()` because `cmd2`.