record mode and full useragent string (#240)
this stashes the mode in the db (live, sandbox, debug). this can be
helpful when doing analyses later down the line in case one needs to
distinguish between live and sandbox assignments in the db. the mode is
stored as part of the taskdata datastring

also, this fixes a neighboring bug where the full useragent string
wasn't actually being saved to the taskdata datastring by backbone