Adding a reference to the LICENSE
I've gone ahead and made a `conda` package out of this repo for distribution via the community-curated `conda-forge` channel. (If y'all want to help maintain it, that would be cool, but is unrelated.) We've started included links to license files in the conda builds themselves, but doing so requires that the license be in the manifest itself. If this pull gets incorporated into the next pypi version, that would let us include said link.