Committing for testing. Bug Fixes and Improvements to Hit Creation
The following bugs and UX issues were noticed when creating a HIT
in the command line tool:

1. Partial command input raises uncaught exception
    Example Input: hit create <numWorkers>
    Now it queries for the remaining arguments.
2. The regex to validate the reward per HIT fails on inputs like:
      -10.00
      3J55
      J55
   New stricter regex used, along with cast to float.
3. Uncaught exception when duration can not be cast as a float, e.g.
   when receiving a non-numeric string as float.

Note, the arguments passed to amt_services methods have been cast
to their numeric equivalents. My tests, and the documentation for
boto.mturk and amazon's amt API suggest this should be just fine.