Retry submission to mturk if failed previously
`or status == SUBMITTED` because we suspect that sometimes the post
to mturk fails after we've set status to SUBMITTED, so really they
have not successfully submitted. This gives another chance for the
submit to work when not using the psiturk ad server.

Note that, unfortunately, we cannot post to turkSubmitTo on behalf
of the turkers in order that we might detect any error response.
MTurk requires that the post originate from the turker's (authenticated)
browser -- and that post cannot happen over AJAX, either :(

A fix for the psiturk ad server will be harder because the ad server
does not understand the notion of 'already_in_db', which is crucial for
this approach to work. `already_in_db == false` + `status == submitted` ensures
that we're looking at a worker that hasn't already completed an assignment
for us. The ad server assumes that anyone with `status > completed` has already
successfully submitted. `/check_worker_status` will need an overhaul for the ad
server to be able to retry posts.