check worker ID during approving and bonusing
Occasionally there are two workers who complete the task for a given Assignment
ID (I think this is caused by certain cases where a worker accidentally logs out
of mTurk), and sometimes there are 0 (if someone submitted without completeing
the task, for example). Currently, "worker approve" and "worker bonus" don't
handle these cases well.

This commit adds functionality to compare the worker id's in the database to the
worker ID that AMT thinks completed the task. It now approves/bonuses the
correct ID when there are multiple, and warns before approving an assignment for
which there is no data in the database.