give server process unique project identifier
Append a short hash of the experiment directory path to the psiturk server
process. This makes it possible to later identify which of several server
processes belongs to the current project. Without this, the same user cannot run
two different psiturk experiments from different folders on different ports.

This solution doesn't allow someone to run two simultaneous servers for the same
project, but I'm not sure why you'd want to do that. It will also break if the
user renames the directory while the server is running.