Updated the content and formating
I've gone through and updated both the content and the formating of this whole document. Most notably, I added that you need to have the location set to "US East" in order to find to the pre-built image at all.

I also really think it should be linked to in the quickstart (see issue #306).