Initial commit of non-functional instructions for setting up psiTurk on Heroku
We've discussed creating docs for setting up psiTurk on Heroku in issue
#306. I've tried following and rewriting some instructions on how to do
this referenced there.

Some of the instructions I couldn't execute without changing the API
calls, and right now, the whole process fails when trying to launch the
app (see comment below).

I'm creating a pull request at this stage to hopefully get some
feedback on how to proceed.