Refactor how server shutdown is accomplished
Existing code used a http request to get the process ID of the server, since the command to shut the server down runs under a different process. This had two consequences:

1. the PID was a publicly accessible via http request
2. the request is asynchronous

The call to kill the server process was also asynchronous (os.kill)

This introduced a race condition that sometimes caused the program to report that the process was still running.

Also, when running a remote shell, the process id of the server could not be idenified using the existing means, if the connection was interrupted and a new shell established.

Changes here:

- uses psutil to iterate through the processes looking for psiturk_experiment_server + a particular known hash and the term 'master'
- uses psutil to recursively kill the process tree
- removes the route
- adds a SIGTERM handler to experiment.py and removes the on_exit hook from experiment_server.py to handle the situation when, within psiturk-shell, ^C is pressed. Otherwise the message was showing when running command `psiturk server off`.