Update dashboard.rst
Updating dashboard.rst to instruct users to also set secret_key as a config variable if they want to enable the dashboard. Before this commit, secret_key is not mentioned as a needed variable, leading to a PsiturkException when you enable the dashboard without secret_key.

add a doc section for the secret key